+++ compiler-core/src/ast/visit.rs
@@ -204,8 +204,9 @@ pub trait Visit<'ast> {
         type_: &'ast Arc<Type>,
         fun: &'ast TypedExpr,
         arguments: &'ast [TypedCallArg],
+        argument_parentheses: &'ast Option<SrcSpan>,
     ) {
-        visit_typed_expr_call(self, location, type_, fun, arguments);
+        visit_typed_expr_call(self, location, type_, fun, arguments, argument_parentheses);
     }
 
     fn visit_typed_expr_bin_op(
@@ -1200,7 +1201,8 @@ where
             type_,
             fun,
             arguments,
-        } => v.visit_typed_expr_call(location, type_, fun, arguments),
+            argument_parentheses,
+        } => v.visit_typed_expr_call(location, type_, fun, arguments, argument_parentheses),
         TypedExpr::BinOp {
             location,
             type_,
@@ -1436,6 +1438,7 @@ pub fn visit_typed_expr_call<'a, V>(
     _type_: &'a Arc<Type>,
     fun: &'a TypedExpr,
     arguments: &'a [TypedCallArg],
+    _arguments_start: &'a Option<SrcSpan>,
 ) where
     V: Visit<'a> + ?Sized,
 {