+++ compiler-core/src/parse.rs
@@ -956,7 +956,7 @@ where
                     }
                 }
                 _ => {
-                    if self.maybe_one(&Token::LeftParen).is_some() {
+                    if let Some((left_paren, _)) = self.maybe_one(&Token::LeftParen) {
                         let start = expr.location().start;
                         match self.maybe_one(&Token::DotDot) {
                             Some((dot_s, _)) => {
@@ -991,7 +991,12 @@ where
                                 // Call
                                 let arguments = self.parse_fn_arguments()?;
                                 let (_, end) = self.expect_one(&Token::RightParen)?;
-                                expr = make_call(expr, arguments, start, end)?;
+                                let argument_parentheses = SrcSpan {
+                                    start: left_paren,
+                                    end,
+                                };
+                                expr =
+                                    make_call(expr, arguments, start, end, argument_parentheses)?;
                             }
                         }
                     } else {
@@ -4898,6 +4903,7 @@ pub fn make_call(
     arguments: Vec<ParserArg>,
     start: u32,
     end: u32,
+    argument_parentheses: SrcSpan,
 ) -> Result<UntypedExpr, ParseError> {
     let mut hole_location = None;
 
@@ -4944,6 +4950,7 @@ pub fn make_call(
         location: SrcSpan { start, end },
         fun: Box::new(fun),
         arguments,
+        argument_parentheses,
     };
 
     match hole_location {